Add files via upload
A hospital readmission basically happens when a patient who is discharged from the hospital, gets re-admitted within a certain period of time. So from the hospital readmission rates we can determine the considered quality of hospital, treatment and the cost of the treatment. So the main objective for this problem is to predict whether a patient is likely to be readmitted to hospital based on the previous details of the patient. The dataset contains 48 features and 66587 rows in train dataset and 16647 data in test dataset. The Column readmission is the target dataset.